+++ src/components/ParticleGenerator.tsx
@@ -1,8 +1,10 @@
 import { useRef } from "react";
 import { useFixedUpdate, useProperty, useUpdate } from "../hooks";
+import { Prop } from "../types";
 import { Particle } from "../utils";
 
 export type ParticleGeneratorProps = {
+  active?: Prop<boolean>;
   rate: number;
   lifespan: number;
   onInit: (particle: Particle) => void;
@@ -14,16 +16,18 @@ export const ParticleGenerator: React.FC<ParticleGeneratorProps> = ({ onInit, on
   const active = useRef<Set<Particle>>(new Set());
   const inactive = useRef<Set<Particle>>(new Set());
 
+  const isActive = useProperty(props.active || false);
   const rate = useProperty(props.rate);
   const lifespan = useProperty(props.lifespan);
 
   useFixedUpdate(rate, () => {
-    if (ref.current) {
+    if (ref.current && isActive.current) {
       if (inactive.current.size > 0) {
         // Reuse an inactive particle.
         const particle = [...inactive.current][0];
         inactive.current.delete(particle);
         active.current.add(particle);
+        particle.node.style.display = 'block';
         onInit(particle);
 
       } else {
@@ -31,6 +35,7 @@ export const ParticleGenerator: React.FC<ParticleGeneratorProps> = ({ onInit, on
         const node = ref.current.insertBefore(document.createElement('div'), ref.current.firstChild);
         const particle = new Particle(node);
         active.current.add(particle);
+        particle.node.style.display = 'block';
         onInit(particle);
       }
     }
@@ -44,6 +49,7 @@ export const ParticleGenerator: React.FC<ParticleGeneratorProps> = ({ onInit, on
         if (particle.age.current >= lifespan.current) {
           active.current.delete(particle);
           inactive.current.add(particle);
+          particle.node.style.display = 'none';
         }
       }
     }